requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
boto3==1.9.57
botocore==1.12.57
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.2
dj-database-url==0.5.0
Django==1.11.15
django-cors-headers==2.4.0
django-extensions==2.1.4
django-filebrowser-no-grappelli==3.7.5
django-grappelli==2.11.1
django-heroku==0.3.1
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-simple-captcha==0.5.10
django-storages==1.7.1
django-tinymce4-lite==1.7.2
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
docutils==0.14
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsmin==2.2.2
MarkupSafe==1.1.0
more-itertools==4.3.0
newrelic==4.8.0.110
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.5
py==1.7.0
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
python-dateutil==2.7.5
pytz==2018.5
requests==2.21.0
s3transfer==0.1.13
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1
